+++ Packages/com.noisecrimestudios.obbreflectionprobe/Samples~/Demo/Common/Scripts/Collections/SceneObjectCollectionController.cs
@@ -1,0 +1,13 @@
+﻿using UnityEngine;
+
+namespace NoiseCrimeStudios.Examples.OBBProjection
+{
+	/// <summary>
+	/// Behaviour that can be used to toggle various scene setups of active transforms/GameObjects.
+	/// </summary>
+	public class SceneObjectCollectionController : MonoBehaviour
+    {
+        public CollectionOfCameraPlacement      cameraCollection = new CollectionOfCameraPlacement();
+        public CollectionOfTransformVisabilty   visabilityCollection = new CollectionOfTransformVisabilty();
+	}
+}