+++ Packages/com.noisecrimestudios.obbreflectionprobe/Samples~/Demo/Common/Scripts/SetRotationOfTarget.cs
@@ -1,0 +1,16 @@
+﻿using UnityEngine;
+
+namespace NoiseCrimeStudios.Examples.OBBProjection
+{
+	public class SetRotationOfTarget : MonoBehaviour
+    {
+        public Transform    target;
+        public Vector3      rotation;
+
+        void OnEnable()
+        {
+            if ( null != target )
+                target.localRotation = Quaternion.Euler( rotation );
+        }
+    }
+}