+++ Packages/com.noisecrimestudios.obbreflectionprobe/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2022-09-21
+Updated package to ensure validation with Unity's Package Validation Tool.
+Updated gitignore.
+Samples - Renamed some scenes to better reflect purpose of scene.
+Samples - Cleaned up directories and scripts.
+Runtime - Additional experimental code for Global shader references.
+
 ## [0.6.0] - 2021-01-06
 Complete rewrite of original code to support both single OOBB per scene and multi OOBB per scene.
 Refactored code and reorganised directory structure.